+++ effectful-core/CHANGELOG.md
@@ -1,4 +1,12 @@
 # effectful-core-2.7.0.0 (2026-??-??)
+* Add the `Input` effect (`Effectful.Input.Dynamic`, `Effectful.Input.Static`,
+  `Effectful.Input.Static.Action` and `Effectful.Labeled.Input`) for access to
+  values.
+* Add the `Output` effect (`Effectful.Output.Dynamic`,
+  `Effectful.Output.Static.Action`, `Effectful.Output.Static.Local.Array`,
+  `Effectful.Output.Static.Local.List`, `Effectful.Output.Static.Shared.Array`,
+  `Effectful.Output.Static.Shared.List` and `Effectful.Labeled.Output`) for
+  accumulation of values.
 * Drop support for GHC < 9.6.
 * Add definitions of `rethrowM` to `MonadThrow` and `catchNoPropagate` to
   `MonadCatch` instances for `Eff` when appropriate (`exceptions` >= 0.10.11).